move PageLog and ModuleLog from app to generic
