+++ app/admin.py
@@ -759,22 +759,6 @@ class FeedbackTypeAdmin(admin.ModelAdmin):
     search_fields =  ("name","org_type","org",)
 
 
-@admin.register(PageLog)
-class PageLogAdmin(admin.ModelAdmin):
-    list_display = ["user", "type", "page", "time"]
-    list_filter = ["type", "time", "platform"]
-    search_fields =  ["user__username", "page"]
-    date_hierarchy = "time"
-
-
-@admin.register(ModuleLog)
-class ModuleLogAdmin(admin.ModelAdmin):
-    list_display = ["user", "type", "page", "module_name", "time"]
-    list_filter = ["type", "module_name", "time", "platform", "page"]
-    search_fields = ["user__username", "page", "module_name"]
-    date_hierarchy = "time"
-
-
 @admin.register(AcademicTag)
 class AcademicTagAdmin(admin.ModelAdmin):
     list_display = ["atype", "tag_content"]
